+++ src/index.js
@@ -9,21 +9,24 @@ app.stage.addChild(elements)
 app.stage.addChild(circleSelection)
 app.stage.addChild(guideNumbers)
 
-var showTimingOptions = require('./listeners/show-timing-modal.js')
-import { saveTimingOptions } from './listeners/save-timing-options.js'
+var showTimingOptions = require('./listeners/menus/show-timing-modal.js')
+import { saveTimingOptions } from './listeners/menus/save-timing-options.js'
 
-var showSongOptions = require('./listeners/show-song-modal.js')
-import { saveSongOptions } from './listeners/save-song-options.js'
+var showSongOptions = require('./listeners/menus/show-song-modal.js')
+import { saveSongOptions } from './listeners/menus/save-song-options.js'
 
-var showMenuOptions = require('./listeners/show-menu-modal.js')
-import { saveMenuOptions } from './listeners/save-menu-options.js'
+var showMenuOptions = require('./listeners/menus/show-menu-modal.js')
+import { saveMenuOptions } from './listeners/menus/save-menu-options.js'
 
 var playShortCut = require('./short-cuts/play.js')
 var adjustPlaybackRate = require('./short-cuts/playback-rate.js')
 
-var slowerButton = require('./listeners/playback-rate.js').slower
-var fasterButton = require('./listeners/playback-rate.js').faster
-var playButton = require('./listeners/play.js')
+var slowerButton = require('./listeners/player/adjust-playback-rate.js').slower
+var fasterButton = require('./listeners/player/adjust-playback-rate.js').faster
+var playButton = require('./listeners/player/play.js')
+var onPaused = require('./listeners/player/on-pause.js')
+
+var seekInput = require('./listeners/player/seek-input.js')
 
 import { beatSelectors } from './short-cuts/beat-select.js'
 
@@ -31,10 +34,6 @@ import { noteCreator } from './short-cuts/note-creator.js'
 
 import { deleteMove } from './moves/delete-move.js'
 
-var seekInput = require('./listeners/seek-input.js')
-
 setBackgroundChart()
 
 setSelectionCircles()
-
-var onPaused = require('./listeners/on-pause.js')
