+++ src/listeners/menus/save-menu-options.js
@@ -1,13 +1,13 @@
-var editor = require('../config/editor.js')
-var player = require('../config/youtube.js')
-import { drawGuideNumbers } from '../stage/guideNumbers/draw-guide-numbers.js'
+var editor = require('../../config/editor.js')
+var player = require('../../config/youtube.js')
+import { drawGuideNumbers } from '../../stage/guideNumbers/draw-guide-numbers.js'
 
 var saveMenuButton = document.getElementById('save-menu-changes')
 var menuModal = document.getElementById('menu-modal')
 var guideNumberSelect = document.getElementById('guide-number-select')
 
 function saveChanges () {
-  
+
   editor.adjustments.numbers = guideNumberSelect.value
 
   if ( player.getDuration() > 0 ) {
