+++ src/listeners/menus/save-timing-options.js
@@ -1,9 +1,9 @@
-var danceChart = require('../../data/dance-chart.js')
-var editor = require('../config/editor.js')
-var player = require('../config/youtube.js')
+var danceChart = require('../../../data/dance-chart.js')
+var editor = require('../../config/editor.js')
+var player = require('../../config/youtube.js')
 
-import { drawGuideNumbers } from '../stage/guideNumbers/draw-guide-numbers.js'
-import { redrawNotes } from '../stage/elements/redraw-notes.js'
+import { drawGuideNumbers } from '../../stage/guideNumbers/draw-guide-numbers.js'
+import { redrawNotes } from '../../stage/elements/redraw-notes.js'
 
 var saveTimingButton = document.getElementById('timing-submitter')
 var timingModal = document.getElementById('timing-modal')
