+++ src/listeners/menus/show-menu-modal.js
@@ -1,6 +1,6 @@
-var danceChart = require('../../data/dance-chart.js')
-var editor = require('../config/editor.js')
-var player = require('../config/youtube.js')
+var danceChart = require('../../../data/dance-chart.js')
+var editor = require('../../config/editor.js')
+var player = require('../../config/youtube.js')
 
 let menuModal = document.getElementById('menu-modal')
 let menuButton = document.getElementById('menu-button')
