+++ src/listeners/menus/show-timing-modal.js
@@ -1,6 +1,6 @@
-var danceChart = require('../../data/dance-chart.js')
-var editor = require('../config/editor.js')
-var player = require('../config/youtube.js')
+var danceChart = require('../../../data/dance-chart.js')
+var editor = require('../../config/editor.js')
+var player = require('../../config/youtube.js')
 
 let timingModal = document.getElementById('timing-modal')
 let timingButton = document.getElementById('timing-button')
