+++ src/listeners/player/adjust-playback-rate.js
@@ -1,5 +1,5 @@
-var player = require('../config/youtube.js')
-var editor = require('../config/editor.js')
+var player = require('../../config/youtube.js')
+var editor = require('../../config/editor.js')
 
 function goSlower () {
   if ( editor.status ) {
