+++ src/listeners/player/on-pause.js
@@ -1,5 +1,5 @@
-var player = require('./../config/youtube.js')
-var songManager = require('./../config/song-manager.js')
+var player = require('../../config/youtube.js')
+var songManager = require('../../config/song-manager.js')
 
 function goToNearestBeat () {
   player.seek(songManager.getNearestBeatTime())
